Sickle
Sickles now only work if the starting block is a Plant,Tall Plant or an Ocean Plant

+++ src/main/java/me/KG20/moreoresinone/Tools/Sickle.java
@@ -23,21 +23,22 @@ public class Sickle extends Item {
             int by = pos.getY();
             int bz = pos.getZ();
 
+            if (world.getBlockState(pos).getMaterial() == Material.PLANTS || world.getBlockState(pos).getMaterial() == Material.TALL_PLANTS || world.getBlockState(pos).getMaterial() == Material.OCEAN_PLANT){
+                for (int x = -1; x < 2; x++) {
+                    for (int z = -1; z < 2; z++) {
 
-            for (int x = -1; x < 2; x++) {
-                for (int z = -1; z < 2; z++) {
+                        BlockPos newBlockPos = new BlockPos(bx + x, by, bz + z);
 
-                    BlockPos newBlockPos = new BlockPos(bx + x, by, bz + z);
-
-                    if (world.getBlockState(newBlockPos).getMaterial() == Material.PLANTS || world.getBlockState(newBlockPos).getMaterial() == Material.TALL_PLANTS || world.getBlockState(newBlockPos).getMaterial() == Material.OCEAN_PLANT) {
+                        if (world.getBlockState(newBlockPos).getMaterial() == Material.PLANTS || world.getBlockState(newBlockPos).getMaterial() == Material.TALL_PLANTS || world.getBlockState(newBlockPos).getMaterial() == Material.OCEAN_PLANT) {
                             world.destroyBlock(newBlockPos, true);
-                    }
+                        }
 
+                    }
                 }
-            }
-            stack.setDamage(stack.getDamage() + 1);
-            if(stack.getDamage() >= stack.getMaxDamage()){
-                stack.shrink(1);
+                stack.setDamage(stack.getDamage() + 1);
+                if(stack.getDamage() >= stack.getMaxDamage()){
+                    stack.shrink(1);
+                }
             }
 
         }
